refactor: split parse to multiple files
