+++ lib/parse/user.ts
@@ -1,0 +1,17 @@
+import { User } from '../types'
+
+export const user = ({
+  socialSecurityNumber,
+  isAuthenticated,
+  userFirstName,
+  userLastName,
+  userEmail,
+  notificationId,
+}: any): User => ({
+  personalNumber: socialSecurityNumber,
+  firstName: userFirstName,
+  lastName: userLastName,
+  email: userEmail,
+  isAuthenticated,
+  notificationId,
+})
